==> 3. modeling.ipynb <==
from sentence_transformers import SentenceTransformer

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, StandardScaler, OneHotEncoder
from keras.callbacks import ReduceLROnPlateau
from keras.models import Sequential
from keras.layers import Dense, Conv1D, MaxPooling1D, Flatten, Dropout, BatchNormalization
# from keras.utils import np_utils, to_categorical
from keras.callbacks import ModelCheckpoint

import np_utils
from keras.utils import to_categorical
import pandas as pd
import numpy as np
import os

import tensorflow as tf
tf.__version__

current_path=os.getcwd().replace("\\", "/").replace("c:", "C:")
data_path=current_path+"/wavFiles/"
model_path=current_path+"/model/"

final_df=pd.read_csv(current_path+'/final_df.csv')

final_df.rename(columns={"filename":"wav_id", "label":"final_label"}, inplace=True)

final_df["final_label"].value_counts()

class text_embedding():
  def __init__(self, model_name):
    self.model_name = model_name

  def fit(self, X, y=None):
        return self

  def transform(self, X):
        embedding_model = SentenceTransformer(self.model_name)
        embedding_vec = embedding_model.encode(X['sentence'])
        X_val = np.concatenate((X.drop(['final_label', 'wav_id', 'sentence'], axis = 1), embedding_vec), axis = 1)
        return X_val

def custom_model(x_train):
  model = Sequential()
  model.add(Conv1D(256, kernel_size=5, strides=1, padding='same', activation='relu', input_shape=(x_train.shape[1], 1)))
  model.add(MaxPooling1D(pool_size=5, strides = 2, padding = 'same'))

  model.add(Conv1D(256, kernel_size=5, strides=1, padding='same', activation='relu'))
  model.add(MaxPooling1D(pool_size=5, strides = 2, padding = 'same'))

  model.add(Conv1D(128, kernel_size=5, strides=1, padding='same', activation='relu'))
  model.add(MaxPooling1D(pool_size=5, strides = 2, padding = 'same'))
  model.add(Dropout(0.2))

  model.add(Conv1D(64, kernel_size=5, strides=1, padding='same', activation='relu'))
  model.add(MaxPooling1D(pool_size=5, strides = 2, padding = 'same'))

  model.add(Flatten())
  model.add(Dense(units=32, activation='relu'))
  model.add(Dropout(0.3))

  model.add(Dense(units=2, activation='softmax'))  # 유닛 수를 2로 변경
  model.compile(optimizer = 'adam', loss = 'categorical_crossentropy', metrics = ['accuracy'])

  # model.summary()
  return model


rlrp = ReduceLROnPlateau(monitor='loss', factor=0.4, verbose=0, patience=2, min_lr=0.0000001) #learning rate 조절
# 선택한 모델만 다시 학습 (텍스트 + 음성)
scaler = StandardScaler()
encoder = OneHotEncoder()

model_name='sentence-transformers/all-mpnet-base-v2'
Y = final_df['final_label'].values
Y = encoder.fit_transform(np.array(Y).reshape(-1,1)).toarray()

txt_embed = text_embedding(model_name = model_name)
X = txt_embed.transform(final_df)

x_train, x_test, y_train, y_test = train_test_split(X, Y, random_state=0, shuffle=True)
x_train = scaler.fit_transform(x_train)
x_test = scaler.transform(x_test)

x_train = np.expand_dims(x_train, axis=2)
x_test = np.expand_dims(x_test, axis=2)
x_train.shape, y_train.shape, x_test.shape, y_test.shape

model = custom_model(x_train)
history=model.fit(x_train, y_train, batch_size=64, epochs=50, validation_data=(x_test, y_test), callbacks=[rlrp])

test_loss, test_acc = model.evaluate(x_test, y_test, verbose=0)
print("Pre-trained Model: ", model_name)
print("Test Accuracy: ",test_acc)

with open(model_path+'text_model_name.txt', 'w') as file:
    file.write(model_name)
import joblib
joblib.dump(encoder, model_path+'encoder.pkl')
joblib.dump(scaler, model_path+'scaler.pkl')
model.save(model_path+'my_model.h5')

final_df
